fix submittion endpint
